moves assigment3 to assignment 1 package
